requirement.txt
fastapi
uvicorn
sqlalchemy
ortools
openpyxl
pydantic
pydantic-settings
aiofiles
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]